feat(us-021): persistence integration tests covering all repository CRUD + BI-012/BI-013
- TestInMemoryConnectionFactory: named shared-memory SQLite (anchor keeps DB alive)
- MonitoredSystemRepository: fixed SetMaintenanceModeAsync to persist MaintenanceOperator
- All DAO records: int -> long for SQLite INTEGER columns (Dapper Int64 mapping)
- 32 integration tests pass: DatabaseInitializer, all 9 repositories, BI-012/BI-013 rules

+++ src/BrokerageMonitor.Infrastructure/Persistence/Repositories/HealthMonitorDefinitionRepository.cs
@@ -215,11 +215,11 @@ public sealed class HealthMonitorDefinitionRepository : IHealthMonitorDefinition
         string  DeadlineTime,
         string  ScheduleType,
         string? CronExpression,
-        int?    DayOfWeek,
+        long?   DayOfWeek,
         string  EmailRecipients,
         string? TeamsWebhookUrl,
-        int     SendOnFailure,
-        int     IsActive);
+        long    SendOnFailure,
+        long    IsActive);
 
     private sealed record JunctionRow(
         string DefinitionId,
